refactor(layer-metadata): move repeated component logic to a composable
