+++ src/components/catalog/catalog-tree.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import { ShallowRef, shallowRef, watchEffect, onMounted } from 'vue'
+import { ShallowRef, shallowRef, watchEffect } from 'vue'
 
 import useLayers from '@/composables/layers/layers.composable'
 import { ThemeNodeModel } from '@/composables/themes/themes.model'
@@ -9,15 +9,16 @@ import LayerTreeNode from '@/components/layer-tree/layer-tree-node.vue'
 import { themesToLayerTree } from '@/components/layer-tree/layer-tree.mapper'
 import type { LayerTreeNodeModel } from '@/components/layer-tree/layer-tree.model'
 import { layerTreeService } from '@/components/layer-tree/layer-tree.service'
-import { layerMetadataService } from '@/composables/layer-metadata/layer-metadata.service'
+import useLayerMetadata from '@/composables/layer-metadata/layer-metadata.composable'
 import LayerMetadata from '../layer-metadata/layer-metadata.vue'
 import { LayerMetadataModel } from '@/composables/layer-metadata/layer-metadata.model'
-import { useTranslation } from 'i18next-vue'
 
 const mapStore = useMapStore()
 const themeStore = useThemeStore()
 const layers = useLayers()
 const layerTree: ShallowRef<LayerTreeNodeModel | undefined> = shallowRef()
+const { metadata, displayLayerMetadata, closeLayerMetadata } =
+  useLayerMetadata()
 
 watchEffect(updateLayerTree)
 
@@ -47,33 +48,12 @@ function toggleParent(node: LayerTreeNodeModel) {
 function toggleLayer(node: LayerTreeNodeModel) {
   layers.toggleLayer(parseInt(node.id, 10), !node.checked)
 }
-
-const { i18next } = useTranslation()
-const metadata: ShallowRef<LayerMetadataModel | undefined> = shallowRef()
-const displayedMetadataNode: ShallowRef<LayerTreeNodeModel | undefined> =
-  shallowRef()
-
-async function displayLayerMetadata(node: LayerTreeNodeModel) {
-  metadata.value = await layerMetadataService.getLayerMetadata(node, i18next)
-  displayedMetadataNode.value = node
-}
-
-onMounted(() => {
-  i18next.on('languageChanged', () => {
-    if (displayedMetadataNode.value)
-      displayLayerMetadata(displayedMetadataNode.value)
-  })
-})
-
-function closeLayerMetadata() {
-  metadata.value = undefined
-}
 </script>
 
 <template>
   <layer-metadata
     v-if="metadata"
-    :layer-metadata="metadata"
+    :layer-metadata="(metadata as LayerMetadataModel)"
     @close-layer-metadata="closeLayerMetadata"
   ></layer-metadata>
   <layer-tree-node
